Create a detail dialog for each item in order to showcase more info.
IMPORTANT:

Currently, there isn't really a way to do a FullScreenDialog with
Compose. The way they've written the Dialog internals always grabs a
default theme* that can't be overwritten. In order to bypas this, I
copied the Dialog implementation in its entirety and changed the
hardcoded theme to be one that is full screen. There was no way to
access the styles of my app from the composable (not one that I could
find easily, anyway), so I hardcoded a singleton int in the Application
object so it was accessible.

*You can override the default dialog theme for the entire app. This
would preculde you from being able to have dialogs that are _not_ full
screen. I wanted to be capable of being selective.